pyusbpd.header: Move header classes into a new file
